+++ karmayogi/src/app/reports/[...campaign]/components/Devices.tsx
@@ -1,0 +1,111 @@
+"use client"
+
+import {TabletSmartphone,MonitorDot } from "lucide-react"
+import { Label, PolarRadiusAxis, RadialBar, RadialBarChart } from "recharts"
+
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card"
+import {
+  ChartConfig,
+  ChartContainer,
+  ChartTooltip,
+  ChartTooltipContent,
+} from "@/components/ui/chart"
+
+const chartConfig = {
+  desktop: {
+    label: "Desktop",
+    color: "hsl(var(--chart-1))",
+  },
+  mobile: {
+    label: "Mobile",
+    color: "hsl(var(--chart-2))",
+  },
+} satisfies ChartConfig
+
+export function Devices({mobile,desktop}:{mobile:number,desktop:number}) {
+  const chartData = [{desktop: desktop, mobile: mobile }]
+  const totalVisitors = desktop + mobile
+
+  return (
+    <Card className="flex flex-col min-w-[320px]">
+      <CardHeader className="items-center pb-0">
+        <CardTitle>Devices</CardTitle>
+      </CardHeader>
+      <CardContent className="flex flex-1 items-center pb-0 ">
+        <ChartContainer
+          config={chartConfig}
+          className="mx-auto aspect-square w-full max-w-[250px]"
+        >
+          <RadialBarChart
+            data={chartData}
+            endAngle={360}
+            innerRadius={70}
+            outerRadius={100}
+          >
+            <ChartTooltip
+              cursor={false}
+              content={<ChartTooltipContent hideLabel />}
+            />
+            <PolarRadiusAxis tick={false} tickLine={false} axisLine={false}>
+              <Label
+                content={({ viewBox }) => {
+                  if (viewBox && "cx" in viewBox && "cy" in viewBox) {
+                    return (
+                      <text x={viewBox.cx} y={viewBox.cy} textAnchor="middle">
+                        <tspan
+                          x={viewBox.cx}
+                          y={(viewBox.cy || 0) - 1}
+                          className="fill-foreground text-2xl font-bold"
+                        >
+                          {totalVisitors.toLocaleString()}
+                        </tspan>
+                        <tspan
+                          x={viewBox.cx}
+                          y={(viewBox.cy || 0) + 20}
+                          className="fill-muted-foreground"
+                        >
+                          Visitors
+                        </tspan>
+                      </text>
+                    )
+                  }
+                }}
+              />
+            </PolarRadiusAxis>
+            <RadialBar
+              dataKey="desktop"
+              stackId="a"
+              cornerRadius={2}
+              fill="var(--color-desktop)"
+              className="stroke-transparent stroke-2"
+            />
+            <RadialBar
+              dataKey="mobile"
+              fill="var(--color-mobile)"
+              stackId="a"
+              cornerRadius={2}
+              className="stroke-transparent stroke-2"
+            />
+          </RadialBarChart>
+        </ChartContainer>
+      </CardContent>
+      <CardFooter className="w-full flex justify-evenly text-white rounded-b-sm bg-black pt-4">
+          <div className="flex gap-4 items-center">
+            <TabletSmartphone  size={24} className="" />
+            <span className=" font-4xl font-bold">{mobile}</span>
+          </div>
+          <div className="flex gap-4 items-center">
+            <MonitorDot size={24} className="" />
+            <span className="font-4xl font-bold">{desktop}</span>
+          </div>
+      </CardFooter>
+    </Card>
+  )
+}